backend/requirements.txt
whitenoise==5.2.0
gunicorn==20.1.0
Django==3.2.4 # High-level Python Web framework - https://github.com/django/django
djangorestframework==3.12.4
pandas==1.2.4
boto3==1.17.78
joblib==1.0.1
django-cors-headers==3.7.0
scikit-learn==0.24.2
covid-data-api==1.5
xgboost==1.4.2
lightgbm==3.2.1

